// agenda-electronique/src/app/models/agenda.model.ts
export class EventModel {
  'title': String;
  'place': String;
  'startDate': String; // C7 de type String car je le stocke des chaines de caractere 
  'endDate': Date;
  'startHour': String;
  'endHour': String;
  'recurrence': String;
  'dateRecurrence': String; // C7 de type String car je le stocke des chaines de caractere 
  'checkboxJourEntier': Boolean;
}